requirements.txt
ultralytics
opencv-python
mediapipe
gradio
numpy
label-studio
pillow